Move DataModelHandler from app/server to app/util
DataModelHandler is used on both controller and accessory
side, so it makes more sense to keep it in the util
directory.
