+++ src/app/server/Server.cpp
@@ -20,9 +20,9 @@
 #include <app/server/Server.h>
 
 #include <app/InteractionModelEngine.h>
-#include <app/server/DataModelHandler.h>
 #include <app/server/EchoHandler.h>
 #include <app/server/RendezvousServer.h>
+#include <app/util/DataModelHandler.h>
 
 #include <ble/BLEEndPoint.h>
 #include <core/CHIPPersistentStorageDelegate.h>
