+++ src/controller/BUILD.gn
@@ -18,7 +18,7 @@ static_library("controller") {
   output_name = "libChipController"
 
   sources = [
-    "${chip_root}/src/app/server/DataModelHandler.cpp",
+    "${chip_root}/src/app/util/DataModelHandler.cpp",
     "${chip_root}/src/app/util/CHIPDeviceCallbacksMgr.cpp",
     "CHIPCluster.cpp",
     "CHIPCluster.h",
